requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
python-jose
pydantic
python-multipart
joblib
numpy
scikit-learn==1.6.1
easyocr
torch
transformers
xgboost
google-genai
geopy